requirements.txt
asgiref==3.2.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.4
django-allauth==0.44.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.3.0
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
django-sslserver==0.22
idna==2.10
oauthlib==3.1.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.7.3
python3-openid==3.2.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
urllib3==1.26.5
XlsxWriter==1.4.3
